# Conflicts:
#	src/Nethermind/Nethermind.Blockchain.Test/BlockTreeTests.cs
#	src/Nethermind/Nethermind.Blockchain/BlockTree.cs
#	src/Nethermind/Nethermind.Blockchain/IBlockTree.cs
#	src/Nethermind/Nethermind.Db/IDbProvider.cs
#	src/Nethermind/Nethermind.JsonRpc.Test/Modules/BoundedModulePoolTests.cs
#	src/Nethermind/Nethermind.JsonRpc.Test/Modules/SingletonModulePoolTests.cs
#	src/Nethermind/Nethermind.Merge.Plugin.Test/EngineModuleTests.HelperFunctions.cs
#	src/Nethermind/Nethermind.Merge.Plugin.Test/EngineModuleTests.Setup.cs
#	src/Nethermind/Nethermind.Merge.Plugin.Test/EngineModuleTests.V1.cs
#	src/Nethermind/Nethermind.Merge.Plugin/Handlers/V1/ForkchoiceUpdatedV1Handler.cs
#	src/Nethermind/Nethermind.Merge.Plugin/MergePlugin.cs

+++ src/Nethermind/Nethermind.Consensus/Processing/BlockchainProcessor.cs
@@ -60,6 +60,7 @@ namespace Nethermind.Consensus.Processing
 
         private int _currentRecoveryQueueSize;
         private readonly CompositeBlockTracer _compositeBlockTracer = new();
+        private Stopwatch _stopwatch = new();
 
         /// <summary>
         /// 
@@ -237,9 +238,7 @@ namespace Nethermind.Consensus.Processing
 
         private void RunProcessingLoop()
         {
-            _stats.Start();
-            if (_logger.IsDebug)
-                _logger.Debug($"Starting block processor - {_blockQueue.Count} blocks waiting in the queue.");
+            if (_logger.IsDebug) _logger.Debug($"Starting block processor - {_blockQueue.Count} blocks waiting in the queue.");
 
             FireProcessingQueueEmpty();
 
@@ -253,6 +252,7 @@ namespace Nethermind.Consensus.Processing
                 Block block = blockRef.Block;
 
                 if (_logger.IsTrace) _logger.Trace($"Processing block {block.ToString(Block.Format.Short)}).");
+                _stats.Start();
 
                 Block processedBlock = Process(block, blockRef.ProcessingOptions, _compositeBlockTracer.GetTracer());
                 if (processedBlock == null)
@@ -313,9 +313,8 @@ namespace Nethermind.Consensus.Processing
             ProcessingBranch processingBranch = PrepareProcessingBranch(suggestedBlock, options);
             PrepareBlocksToProcess(suggestedBlock, options, processingBranch);
 
-            Stopwatch stopwatch = Stopwatch.StartNew();
+            _stopwatch.Restart();
             Block[]? processedBlocks = ProcessBranch(processingBranch, suggestedBlock, options, tracer);
-            stopwatch.Stop();
             if (processedBlocks == null)
             {
                 return null;
@@ -338,19 +337,22 @@ namespace Nethermind.Consensus.Processing
 
             if ((options & (ProcessingOptions.ReadOnlyChain | ProcessingOptions.DoNotUpdateHead)) == 0)
             {
-                if (lastProcessed!.IsPostMerge == false)
-                {
-                    _blockTree.UpdateMainChain(processingBranch.Blocks.ToArray(), true);
-                }
-                else
-                {
-                    if (_logger.IsTrace)
-                        _logger.Trace(
-                            $"Marked blocks as processed {lastProcessed}, blocks count: {processedBlocks.Length}");
-                    _blockTree.MarkChainAsProcessed(processingBranch.Blocks.ToArray());
-                }
+                if (_logger.IsTrace)
+                    _logger.Trace(
+                        $"Updating main chain: {lastProcessed}, blocks count: {processedBlocks.Length}");
+                _blockTree.UpdateMainChain(processingBranch.Blocks.ToArray(), true);
 
-                Metrics.LastBlockProcessingTimeInMs = stopwatch.ElapsedMilliseconds;
+                Metrics.LastBlockProcessingTimeInMs = _stopwatch.ElapsedMilliseconds;
+            }
+
+            if ((options & ProcessingOptions.MarkAsProcessed) != 0)
+            {
+                if (_logger.IsTrace)
+                    _logger.Trace(
+                        $"Marked blocks as processed {lastProcessed}, blocks count: {processedBlocks.Length}");
+                _blockTree.MarkChainAsProcessed(processingBranch.Blocks.ToArray());
+
+                Metrics.LastBlockProcessingTimeInMs = _stopwatch.ElapsedMilliseconds;
             }
 
             if ((options & ProcessingOptions.ReadOnlyChain) == ProcessingOptions.None)
@@ -452,8 +454,8 @@ namespace Nethermind.Consensus.Processing
 
             finally
             {
-                if (invalidBlockHash is not null)
-                { 
+                if (invalidBlockHash is not null && (options & ProcessingOptions.ReadOnlyChain) == 0)
+                {
                     DeleteInvalidBlocks(invalidBlockHash);
                 }
             }
@@ -573,7 +575,7 @@ namespace Nethermind.Consensus.Processing
                                                         branchingPoint.Hash) == false;
                 if (_logger.IsTrace)
                     _logger.Trace(
-                        $" Processing conditions notFoundTheBranchingPointYet {notFoundTheBranchingPointYet}, notReachedTheReorgBoundary: {notReachedTheReorgBoundary}, suggestedBlockIsPostMerge {suggestedBlockIsPostMerge}, postMergeFinishBranchingCondition: {postMergeFinishBranchingCondition}");
+                        $" Current branching point: {branchingPoint.Number}, {branchingPoint.Hash} TD: {branchingPoint.TotalDifficulty} Processing conditions notFoundTheBranchingPointYet {notFoundTheBranchingPointYet}, notReachedTheReorgBoundary: {notReachedTheReorgBoundary}, suggestedBlockIsPostMerge {suggestedBlockIsPostMerge}, postMergeFinishBranchingCondition: {postMergeFinishBranchingCondition}");
             } while (preMergeFinishBranchingCondition || postMergeFinishBranchingCondition);
 
             if (branchingPoint != null && branchingPoint.Hash != _blockTree.Head?.Hash)
